Updates to licenses, mailing list and added release manager pgp keys.
The license block was outdated. Updated it to point to the Apache 2.0
license as the main license.

The Mailing list heading was still pointing to the old mailing list.
Updated to point to discourse.

Added new heading with the PGP signing keys. Added Tobias's and Tom's
keys to the site.